+++ mk/arch/arm64/include/vmm.h
@@ -96,10 +96,10 @@ void  kfree(void *addr);
 void *vmm_temp_map(uint64_t phys);
 void  vmm_temp_unmap(void);
 
-int copy_from_user(void *dst, const void *user_src, uint32_t size);
-int copy_to_user(void *user_dst, const void *src, uint32_t size);
-int user_range_ok(const void *uaddr, uint32_t size, int write);
-int strncpy_from_user(char *dst, const char *user_src, uint32_t max_len);
+int copy_from_user(void *dst, const void *user_src, size_t size);
+int copy_to_user(void *user_dst, const void *src, size_t size);
+int user_range_ok(const void *uaddr, size_t size, int write);
+int strncpy_from_user(char *dst, const char *user_src, size_t max_len);
 
 int vmm_handle_page_fault(registers_t *r, uint64_t fault_addr, uint32_t esr);
